Rearranged python code
